Libraries updates and cleanup
 * Updating common/lib
 * Updating lib/csu
 * Updating lib/libc
 * Updating libexec/ld.elf_so
 * Corrected test on __minix in featuretest to actually follow the
   meaning of the comment.
 * Cleaned up _REENTRANT-related defintions.
 * Disabled -D_REENTRANT for libfetch
 * Removing some unneeded __NBSD_LIBC defines and tests

+++ include/ssp/unistd.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: unistd.h,v 1.5 2011/01/19 19:21:29 christos Exp $	*/
+/*	$NetBSD: unistd.h,v 1.6 2011/01/26 18:07:44 christos Exp $	*/
 
 /*-
  * Copyright (c) 2006 The NetBSD Foundation, Inc.
@@ -39,7 +39,7 @@ __BEGIN_DECLS
 __ssp_redirect0(ssize_t, read, (int __fd, void *__buf, size_t __len), \
     (__fd, __buf, __len));
 
-__ssp_redirect(int, readlink, (const char *__restrict __path, \
+__ssp_redirect(ssize_t, readlink, (const char *__restrict __path, \
     char *__restrict __buf, size_t __len), (__path, __buf, __len));
 
 __ssp_redirect(char *, getcwd, (char *__buf, size_t __len), (__buf, __len));
